制卡以及bug修改

+++ src/pages/teaching/device.vue
@@ -1,48 +1,38 @@
 <template>
     <section>
         <!--工具条-->
-        <el-col :span="24" class="toolbar">
-            <el-form :model="deviceCondition" label-width="80px">
-                <el-col :span="6">
-                    <el-form-item label="设备类型" class="normal">
-                        <el-select placeholder="请选择设备类型" v-model="deviceCondition.type" @change="typeChange">
-                            <el-option label="全部" value="1"></el-option>
-                            <el-option label="设备" value="device"></el-option>
-                            <el-option label="摄像头" value="camera"></el-option>
-                        </el-select>
-                    </el-form-item>
-                </el-col>
-                <el-col :span="6">
-                    <el-form-item label="绑定教室" class="normal">
-                        <el-select placeholder="请绑定教室" v-model="deviceCondition.used" @change="usedChange">
-                            <el-option label="全部" value="1"></el-option>
-                            <el-option label="已绑定" value="true"></el-option>
-                            <el-option label="未绑定" value="false"></el-option>
-                        </el-select>
-                    </el-form-item>
-                </el-col>
-                <el-col :span="9">
-                    <el-form-item label="设备编号" class="normal">
-                        <el-input style="width:250px;" icon="search" placeholder="请输入设备编号" v-model="deviceCondition.number" :on-icon-click="queryDeviceList" @keyup.enter.native="queryDeviceList">
-                        </el-input>
-                    </el-form-item>
-                </el-col>
-                <el-col :span="8" style="float:right;">
-                    <el-form-item>
-                        <el-button type="primary" @click="queryDeviceList"> 查 询 </el-button>
-                        <el-button type="primary" @click="addDeviceClick"> 新 增 </el-button>
-                    </el-form-item>
-                </el-col>
+        <el-row class="toolbar">
+            <el-form :model="deviceCondition" :inline="true">
+                <el-form-item class="normal">
+                    <el-select placeholder="请选择设备类型" v-model="deviceCondition.type" @change="typeChange">
+                        <el-option label="全部" value="1"></el-option>
+                        <el-option label="设备" value="device"></el-option>
+                        <el-option label="摄像头" value="camera"></el-option>
+                    </el-select>
+                </el-form-item>
+                <el-form-item class="normal">
+                    <el-select placeholder="请绑定教室" v-model="deviceCondition.used" @change="usedChange">
+                        <el-option label="全部" value="1"></el-option>
+                        <el-option label="已绑定" value="true"></el-option>
+                        <el-option label="未绑定" value="false"></el-option>
+                    </el-select>
+                </el-form-item>
+                <el-form-item class="normal">
+                    <el-input placeholder="请输入设备编号" v-model="deviceCondition.number" style="width:260px;">
+                    </el-input>
+                </el-form-item>
+                <el-button type="primary" @click="queryDeviceList"> 查 询 </el-button>
+                <el-button type="primary" @click="addDeviceClick" class="right"> 新 增 </el-button>
             </el-form>
-        </el-col>
+        </el-row>
         <!--列表-->
-        <el-table highlight-current-row :data="deviceList" highlight-current-row>
+        <el-table :data="deviceList">
             <el-table-column prop="number" label="设备编码"></el-table-column>
-            <el-table-column prop="type" label="设备类型"></el-table-column>
+            <el-table-column prop="type" label="设备类型" :formatter="formatter"></el-table-column>
             <el-table-column prop="model" label="设备型号"></el-table-column>
             <el-table-column prop="manufacturer" label="厂商"></el-table-column>
-            <el-table-column prop="used" label="绑定教室"></el-table-column>
-            <el-table-column prop="gmtCreate" label="创建时间"></el-table-column>
+            <el-table-column prop="bind" label="是否绑定教室" :formatter="formatter"></el-table-column>
+            <el-table-column prop="gmtCreate" label="创建时间" :formatter="formatter"></el-table-column>
             <el-table-column prop="task" label="操作">
                 <template scope="scope">
                     <el-button type="text" size="small" @click="updateDeviceClick(scope.row)">编辑</el-button>
@@ -51,29 +41,29 @@
             </el-table-column>
         </el-table>
         <!--工具条-->
-        <el-pagination layout="total, prev, pager, next, jumper" @current-change="handleCurrentChange" :page-size="pageSize" :total="this.total">
+        <el-pagination layout="total, prev, pager, next, jumper" @current-change="handleCurrentChange" :page-size="pageSize" :total="total">
         </el-pagination>
         <!--新增设备-->
         <el-dialog title="新增设备" v-model="addDevice" :close-on-click-modal="false" size="tiny">
             <el-row style="margin-top:60px">
-                <el-form :rules="validate" :model="addDeviceCondition" label-width="70px" ref="addDeviceRef">
+                <el-form :rules="validate" :model="addDeviceCondition" label-width="100px" ref="addDeviceRef">
                     <el-col :span="0">
-                        <el-form-item label="设备类型" class="normal" prop="type">
-                            <el-select placeholder="请选择设备类型" v-model="addDeviceCondition.type">
+                        <el-form-item label="设备类型" prop="type">
+                            <el-select placeholder="请选择设备类型" v-model="addDeviceCondition.type" style="width:260px;">
                                 <el-option label="设备" value="device"></el-option>
                                 <el-option label="摄像头" value="camera"></el-option>
                             </el-select>
                         </el-form-item>
                         <el-form-item label="编号" prop="number">
-                            <el-input style="width:250px;" v-model="addDeviceCondition.number" placeholder="请输入编号">
+                            <el-input style="width:260px;" v-model="addDeviceCondition.number" placeholder="请输入编号">
                             </el-input>
                         </el-form-item>
                         <el-form-item label="型号">
-                            <el-input style="width:250px;" v-model="addDeviceCondition.model" placeholder="请输入型号">
+                            <el-input style="width:260px;" v-model="addDeviceCondition.model" placeholder="请输入型号">
                             </el-input>
                         </el-form-item>
                         <el-form-item label="生产厂家">
-                            <el-input style="width:250px;" v-model="addDeviceCondition.manufacturer" placeholder="请输入生产厂家">
+                            <el-input style="width:260px;" v-model="addDeviceCondition.manufacturer" placeholder="请输入生产厂家">
                             </el-input>
                         </el-form-item>
                     </el-col>
@@ -86,12 +76,12 @@
         </el-dialog>
         <!--编辑设备-->
         <el-dialog title="编辑设备" v-model="updateDevice" :close-on-click-modal="false" size="tiny">
-             <el-row style="margin-top:33px">
+            <el-row class="mt20">
                 <el-col>
-                    <label style="font-size:17px;">设备类型：{{this.updateDeviceCondition.type}}</label>
+                    <label>设备类型：{{this.updateDeviceCondition.typeName}}</label>
                 </el-col>
-             </el-row>
-            <el-row style="margin-top:40px">
+            </el-row>
+            <el-row class="mt20">
                 <el-form :rules="validate" :model="updateDeviceCondition" label-width="70px" ref="updeatDeviceRef">
                     <el-col :span="0">
                         <el-form-item label="编号" prop="number">
@@ -163,19 +153,25 @@ export default {
         }
     },
     methods: {
+        formatter(row, column) {
+            if (column.property === "bind") {
+                return row.bind ? "绑定" : "未绑定";
+            }
+            else if (column.property === "gmtCreate") {
+                return new Date(row.gmtCreate).Format("yyyy-MM-dd HH:mm:ss");
+            }
+            else if (column.property === "type") {
+                return row.type === "device" ? "设备" : "摄像头";
+            }
+        },
         //查询服务管理列表
         queryDeviceList() {
             let args = [this.schoolCode, this.deviceCondition.used, this.deviceCondition.type, this.deviceCondition.number, this.currentPage, this.pageSize];
-            // console.log("1111111!!!"+args)
-            this.deviceList=[];
             request.teaching.device.query.queryDeviceList(args).then((res) => {
                 if (res.success === true) {
                     this.total = res.object.num;
                     this.deviceList = res.object.list;
-                    for (let i = 0; i < this.deviceList.length; i++) {
-                        this.deviceList[i].type = this.changStatus(this.deviceList[i].type);
-                        this.deviceList[i].gmtCreate = this.changDate(this.deviceList[i].gmtCreate);
-                    }
+                    console.log(this.deviceList);
                 }
             });
         },
@@ -184,7 +180,6 @@ export default {
             this.$refs["addDeviceRef"].validate((valid) => {
                 if (valid) {
                     let args = this.addDeviceCondition;
-                    // console.log("1111111!!!"+args)
                     request.teaching.device.save.saveDevice(args).then((res) => {
                         if (res.success === true) {
                             this.addDevice = false;
@@ -200,25 +195,31 @@ export default {
         //删除按钮
         deleteDevice(row) {
             let args = [row.id];
-            // console.log("1111111!!!"+args)
-            request.teaching.device.delete.deleteDevice(args).then((res) => {
-                if (res.success === true) {
-                    this.queryDeviceList();
-                    this.$message({ message: (res.object != null ? res.object : "") + "删除设备成功！", type: "success" });
-                } else {
-                    this.$message.error("删除设备失败！" + res.message);
-                }
-            });
+            this.$confirm("您正在执行删除设备操作, 是否继续？", "提示", {
+                confirmButtonText: "确定",
+                cancelButtonText: "取消",
+                closeOnClickModal: false,
+                type: "warning"
+            }).then(() => {
+                request.teaching.device.delete.deleteDevice(args).then((res) => {
+                    if (res.success === true) {
+                        this.queryDeviceList();
+                        this.$message({ message: (res.object != null ? res.object : "") + "删除设备成功！", type: "success" });
+                    } else {
+                        this.$message.error("删除设备失败！" + res.message);
+                    }
+                });
+            }).catch(() => { });
         },
         //编辑按钮
         updeatDevice() {
             this.$refs["updeatDeviceRef"].validate((valid) => {
                 if (valid) {
-                    let args ={};
-                    args.id =  this.updateDeviceCondition.id;
-                     args.number =  this.updateDeviceCondition.number;
-                      args.model =  this.updateDeviceCondition.model;
-                       args.manufacturer =  this.updateDeviceCondition.manufacturer;
+                    let args = {};
+                    args.id = this.updateDeviceCondition.id;
+                    args.number = this.updateDeviceCondition.number;
+                    args.model = this.updateDeviceCondition.model;
+                    args.manufacturer = this.updateDeviceCondition.manufacturer;
                     // console.log("1111111!!!"+args)
                     request.teaching.device.save.updeatDevice(args).then((res) => {
                         if (res.success === true) {
@@ -285,14 +286,14 @@ export default {
             this.updateDeviceCondition.number = row.number;
             this.updateDeviceCondition.model = row.model;
             this.updateDeviceCondition.manufacturer = row.manufacturer;
+            this.updateDeviceCondition.typeName = row.type === "device" ? "设备" : "摄像头";
         }
     },
+    //页面激活之后
+    activated() {
+        this.queryDeviceList();
+    },
     mounted() {
-        this.queryDeviceList();
-    },
-     //页面激活之后
-    activated(){
-        this.queryDeviceList();
-    },
+    }
 }
 </script>
